Final Version of the program.
GUI will be implemented later. 
This program gives you the login and logout time of the student in terms of the status key.
The same can be implemented to track employees in a company.
The database connection should be configured to your requirements. Xampp was used to host a local database.